refactor/ token util package로 이동
