+++ src/utils/index.ts
@@ -1,3 +1,2 @@
 export * from './types';
-export * as token from './token';
 export * from './context';
